Fix release:prepare not committing the version into the tag
We just encountered the problem where release:prepare would tag the
current commit -- without adjusting the <version> tag at all!

Apparently, this is caused by a faulty maven-scm-provider-gitexe version
referenced in maven-release-plugin version 2.4.2:

	http://stackoverflow.com/a/20657721

Unfortunately, forcing the maven-scm-provider-gitexe version to 1.8.1
did not fix the problem for this developer, but bumping it to 1.9 did.

Since scijava-scripts' release-version.sh relies on release:prepare, we
need to deploy this change right away.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>