Make pom-scijava more useful as a parent POM
We will want to deploy more and more things to OSS Sonatype, just like
pom-scijava itself.

As such, it would be not quite bright to insist on repeating the same
configurations over and over again when Maven allows you to inherit
from parent POMs!

To help this, we do two things: first, we explicitly remove the
explicit statement asking *not* to inherit the configuration. This
was stupidly copied from some website by this developer, so it is
only fair that this developer fixes it.

Second, we add the explicit configuration of the maven-source-plugin
and the maven-javadoc-plugin, as per OSS Sonatype's requirements.
That way, all child projects will automatically attach all the correct
secondary artifacts when the profile 'sonatype-oss-release' is active.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>