Release version 2.0
This removes the groupId properties, bumps scijava-common and
scijava-ui-swing, and removes the ImageJ managed dependencies
(they migrated to pom-imagej).

If this approach works well, we'll migrate the ImgLib2 and SCIFIO
managed dependencies to their own parent POMs as well. In this way,
the pom-imglib2, pom-scifio and pom-imagej POMs can serve as examples
for how other projects can extend the SciJava POM as they see fit.