Add a profile to ease debugging via remote debuggers
In quite a few cases, unit tests succeed in interactive debuggers
such as Eclipse's but fail on the command-line. Or on a different
computer, e.g. on a Jenkins node.

The new 'debug-port' profile -- conveniently activated by setting the
'debugPort' property -- helps with that:

	mvn -DdebugPort=8000,suspend=y -Dtest=MyFailingTest

will ask Maven to prepare everything to run MyFailingTest (which can be
in a package, you do not have to specify that in the 'test' property)
and then wait for a remote debugger to attach.

In Eclipse, you can easily configure such a remote debugger: after
launching Run>Debug Configurations..., click on 'Remote Java
Application' and then hit the 'New' button (or the context menu).
Eclipse defaults to TCP port 8000 -- hence the example above -- so you
can easily port-forward via ssh if needed.

See also http://loci.wisc.edu/software/debugging-techniques

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>