Simplify generation of unit tests JARs
This updates the maven-jar-plugin version to 2.5, sets the skipIfEmpty
property to true, and removes the test-jar profile (previously triggered
on the existence of a src/test/java folder) in favor of always executing
the test-jar goal.

Before setting the skipIfEmpty property, such a configuration would
result in a bogus empty tests JAR being generated for any project
without unit tests. But with skipIfEmpty set, we avoid the problem in a
simpler way that does not require a dedicated profile.

The upgrade to 2.5 is ancillary, but probably doesn't hurt, because
there were some minor skipIfEmpty bugs fixed (MJAR-162, MJAR-167).