Update to 2.0.4 version of Pivot javadocs
The reason is that the 2.0.3 version has gone offline. Unfortunately,
Pivot's policy combines the worst of both worlds: the URL has the
version embedded within it, but then when they do a new release they
take the old version offline. So the URL for the javadocs is
continuously changing.

At some point we'll have to find a more robust way to manage javadoc
linkages...