Add an Eclipse-specific profile to help with annotation processing
Eclipse does not execute annotation processors as part of the build, in
particular not as part of the incremental build. However, scijava-common's
annotation processor is quite able to run as part of the incremental
build.

Let's use pom-scijava to auto-activate the newly introduced goal of the
scijava-maven-plugin to beat Eclipse into submission, so that we can work
around Eclipse's failure to abide by the Java specification (that
*mandates* running all annotation processors available on the class path
*unless* turned off explicitly).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>