Add maven-enforcer plugin
Adds configuration specification for the maven-enforcer plugin and a
default binding to the validate lifecycle phase. This will cause the
maven-enforcer:enforce target to be run at the start of each maven
build, using whichever rules are active.

As configured, these rules are always on:
* enforceBytecodeVersion
* banCircularDependencies

These rules are on by default but can be disabled by setting the system
property: scijava.enforce.skip:
* requireReleaseDebs
* banDuplicateClasses
* requireMavenVersion
* requirePrerequisites
* requireReleaseDeps

As some of these properties may use maven and jvm versions, these have
been converted to their own properties.