Special-case JRuby/Jython in the maven-enforcer-plugin configuration
As of pom-scijava 3.7, the maven-enforcer-plugin runs with every Maven run
by default. Unfortunately, neither JRuby nor Jython dependencies abide by
the rules we want to enforce, nor is it within the time constraints of
the SciJava project to try to get it fixed upstream. Therefore, let's
just work around the problematic parts in those dependencies.

The original plan was to put this into the pom.xml files of
scripting-jython and scripting-jruby, respectively. However, the
maven-enforcer-plugin ignores the configuration of dependencies' POMs,
instead only looking at the current project POM and its ancestors.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>