Silence annoying Eclipse-related warning
When running certain Maven commands from the CLI (such as
"mvn tidy:pom"), the following annoying warning is printed:

[WARNING] The POM for org.eclipse.m2e:lifecycle-mapping:jar:1.0.0 is missing, no dependency information available
[WARNING] Failed to retrieve plugin descriptor for org.eclipse.m2e:lifecycle-mapping:1.0.0: Plugin org.eclipse.m2e:lifecycle-mapping:1.0.0 or one of its dependencies could not be resolved: Could not find artifact org.eclipse.m2e:lifecycle-mapping:jar:1.0.0 in central (http://repo.maven.apache.org/maven2)

We avoid the problem by wrapping the lifecycle-mapping plugin
configuration inside of a profile, which is only active when working
with the project within Eclipse.

Thanks to Johannes Schindelin for discovering this hack.

See also:
http://stackoverflow.com/a/23707050/1207769