Bump to next major version: 6.0.0-SNAPSHOT
This marks the BOM as backwards-incompatible. This should have been done
as part of 6264e74bdcb396682d1102b1aa0c76b35ea17791. I was somehow
thinking that the bump of scijava-config to 2.0.0 would not be backwards
incompatible from a BOM perspective. However, my thinking was fuzzy and
I cannot really justify it clearly -- better to be consistent and follow
the rule that minor dependency version bump(s) yield a minor BOM version
bump, and ditto for major version bumps.