Do not fork the lifecycle when attaching sources
In general, it is better to use jar-no-fork than jar for a goal.
And the latest edition of the OSS Sonatype guide agrees with this
change, too:
http://central.sonatype.org/pages/apache-maven.html#javadoc-and-sources-attachments

Note that for now, we do _not_ update the maven-javadoc-plugin's usage
of <goal>jar</goal>, because the Sonatype guide still uses jar for that.