Name version properties consistently
They should _always_ be of the form artifactId.version where artifactId
is the specific dependency's artifactId. This will be very helpful to
allow the melting pot to properly override all dependency versions.

To avoid skew between multiple components versioned in synchrony, we
still keep a master version property for the multi-component project as
a whole, and declare the individual component properties to equal that
master version property.

Because this renames some version properties, it bumps major version.