Add Eclipse config for build-helper-maven-plugin
On out-of-the-box Eclipse installations, there is no m2e connector for
the build-helper-maven-plugin. It used to be the case (I think) that you
could simply use the quick fix solution to install the m2e connector,
but these days it is not available from the Eclipse marketplace (at
least not by default).

Regardless, it's better to simply provide the lifecycle mappings in
the POM configuration directly, so everything "just works."