Run the enforcer in Eclipse only on configuration
It is intended to check the structure of the dependencies.
Changes to source code will not affect whether it succeeds.
So it only needs to run when the POM is changed.

This avoids an issue where complex projects such as Fiji have hiccups
during incremental builds, due to lots of dependency validation.