Remove no-longer-needed exclusion rules
For JAMA, everything in the SciJava software stack now uses the
official gov.nist.math:jama artifact, rather than the jama:jama
we rolled ourselves. So no more need to worry about duplication.

For tools, that is a system scope dependency of ImageJ 1.x,
for compilation purposes only. It should really be declared as
"provided + system" but you cannot do that with Maven.
In any case, we do not want anything to inherit it transitively.
So the proper fix will be an exclusion at the pom-imagej level.