Remove remaining BanDuplicateClasses exceptions
I tested with the latest fiji/fiji, and the enforcer checks all pass.
So I guess (hope!) that things are simply cleaner now than in the past,
when these exceptions were introduced!

For sanity, I also tested scijava/scripting-jruby (since one big block
of exceptions was for JRuby in particular), and it also passes.

So, we'll cut a release of pom-scijava without any enforcer exceptions.
And if downstream projects have trouble as a consequence when updating
to the new pom-scijava parent, we can cut another release that restores
exceptions as needed.