Define main-class and package-name JAR properties
Now, to add a Main-Class to the JAR manifest, downstream POMs can simply
override the main-class property. And same for Package-Name with the
package-name property.

This has the wonderful side effect of eliminating the
following obnoxious new warning in Eclipse Neon:

    You have to use a classifier to attach artifacts to the project instead of replacing them.
    (org.apache.maven.plugins:maven-jar-plugin:3.0.2:jar:default-jar:package)

This message is shown as an error on the pom.xml in Eclipse Neon
whenever a <plugin> block for maven-jar-plugin appears downstream for
the purpose of adding a Main-Class and/or Package-Name.