Downgrade maven-jar-plugin to version 2.6
This avoids a bug in the M2E plugin which ships with Eclipse Neon,
which issues the following warning:

    You have to use a classifier to attach artifacts to the project instead of replacing them.
    (org.apache.maven.plugins:maven-jar-plugin:3.0.2:jar:default-jar:package)

I thought I had found a workaround with 85f49450, but in actuality
it does not work after refreshing the project.

It seems like this issue was addressed in the M2E source code:

    https://github.com/tesla/m2eclipse-mavenarchiver/issues/8

However, installing the latest m2eclipse-mavenarchiver from the
update site mentioned in the comment thread does not address the issue
in my tests. And even if it did, this would only avoid the problem for
people in their local installs. Rather than require everyone to do
that, let's instead downgrade to a version of maven-jar-plugin which
works fine. We can upgrade it to the latest again once the bug is
fixed and ships with the newest Eclipse download.