Update Interactive_3D_Surface_Plot to 3.0.0
The 2.31.2 version kept being detected as newer than 2.4.x -- and
arguably rightly so, since it is a later version by SemVer.

To avoid this conundrum, we bump to 3.0.0 which is clearly newer
than both. This is _not_ actually a breaking change in the API
though, so let's not do another major pom-scijava version bump.