Downgrade javassist to 3.22.0-GA
The 3.23.1-GA version has a bug preventing successful generation of
a CannotCompileException; instead, a NullPointerException is thrown.

This is already reported as jboss-javassist/javassist#209, and fixed in
jboss-javassist/javassist@a3d1aa25133c2ec473e9b3ebe2bfa5b4ebe5759e.

But no new release yet as of this writing, so let's stick with the
known-to-work version for the moment.