Add exclusions needed for Apache Spark
These exclusions avoid duplicate classes between spark-core_2.11
and other components of the SciJava component collection.

There is one conflict remaining, with these components:

* ome:formats-bsd
* ome:formats-gpl
* org.openmicroscopy:ome-common

The conflict is between:

* com.esotericsoftware:kryo-shaded:3.0.2 (dependency of spark-core_2.11)
* com.esotericsoftware.kryo:kryo:2.24.0 (dependency of artifacts above)

Downstream projects wishing to depend on both spark-core_2.11 and
one or more of those Bio-Formats artifacts will need to choose which
kryo version to keep on the classpath.

For example, to keep kryo 2.24.0, the following exclusion block can be
added to the spark-core_2.11 dependency declaration:

  <exclusions>
    <exclusion>
      <!-- Conflicts with com.esotericsoftware.kryo:kryo -->
      <groupId>com.esotericsoftware</groupId>
      <artifactId>kryo-shaded</artifactId>
    </exclusion>
  </exclusions>

But as of this writing, I do not know whether kryo 2.24.0
is compatible with Spark's Kryo-related functionality.