Update imagej-common to 0.28.0
This removes the obsolete net.imagej.table API.
It migrated to org.scijava.table in scijava-table.

Leaving it in place but deprecated caused skew across the
org.scijava.table.Table and net.imagej.table.Table interfaces; e.g.:

.../imagej-legacy/src/main/java/net/imagej/legacy/translate/ResultsTableHarmonizer.java:[137,53] no suitable method found for add(net.imagej.table.ResultsTable)
    method java.util.Collection.add(net.imagej.table.Table<?,?>) is not applicable
      (argument mismatch; net.imagej.table.ResultsTable cannot be converted to net.imagej.table.Table<?,?>)
    method java.util.List.add(net.imagej.table.Table<?,?>) is not applicable
      (argument mismatch; net.imagej.table.ResultsTable cannot be converted to net.imagej.table.Table<?,?>)

Better to just remove the old API and let the compiler report directly
when we are still using it.