Declare ij.version property for ImageJ1
Otherwise, the melting pot's logic does not override net.imagej:ij.

It will still be an issue for older component releases with a dependency
on ImageJ1; the workaround on the melting pot side is to manually add an
override in build.sh for the older imagej1.version property as well.