Move N5 into the SciJava component collection
The N5 project is now part of the SciJava component collection,
meaning it will be tested in the melting pot to help validate
that the SciJava Bill of Materials works with N5 as intended.

N5's dependencies remain in the third party components section,
and are now sorted into the list by project name.