Exclude old artifactIds in favor of new ones
* javax.annotation:jsr250-api -> javax.annotation:javax.annotation-api
* javax.transaction:jta -> javax.transaction:javax.transaction-api

Unfortunately, this may result in runtime issues if the artifacts in
question (jetty-jspc-maven-plugin and omero-model, respectively) are
depended upon without the newer artifacts on the classpath as well.
But Maven does not have a built-in way to add dependencies to a managed
artifact, so what can you do? Maybe we add a smart dependency checking
step to scijava-maven-plugin the future to catch these sorts of things?