Ignore jruby-core + jnr-ffi conflicts
I don't like doing this, but jruby-core ships jnr-ffi unshaded, and
jruby-core seems to be the most granular artifact available. :-(

Since jython-slim depends properly on jnr-ffi, in order to support
both jruby-core and jython-slim on the same classpath, we need to
ignore this class overlap and just hope for the best version-wise.