mega-melt: explicitly skip ignored artifacts
The SciJava component collection includes some artifacts we do not want
to include in the mega melt, even though their groupIds as a whole are
included. For example, we do not need to melt SciJava forks of
third-party components (Java 3D, JUnit Benchmarks, JEP), and we do not
want to melt known-to-fail components (Scenery, n5-blosc and n5-zarr).