mega-melt: avoid duplicate version properties
When two components have the same artifactId, the melting pot sets two
version properties with the same name, one for each component, likely at
different versions. This breaks the melting pot. To avoid the issue for
now, we prune the known-duplicate properties from the build script.
Some day, it would be nice for the melting pot to use "fully qualified"
version properties that include the groupId somehow, to avoid ambiguity.