Restore imagej.version property
We need a version property matching the artifactId of the artifact,
which in this case is net.imagej:imagej, for the melting pot to work.

I considered renaming net.imagej:imagej to net.imagej:imagej2, but
that would disrupt a large ecosystem of downstream code, including
e.g. PyImageJ, which defaults to using net.imagej:imagej currently.
It would be possible to do this rename later, but it would be more
involved, and I would prefer to do it at a time when the rest of the
system is not in motion so much.