mega-melt: do not include the original ImageJ
ImageJ requires Java 9+ to compile, but the rest of the components
require Java 8 with JavaFX. There is no single Java compiler that can be
used in both scenarios. Anyway, the original ImageJ has no dependencies,
so no changes to version pins will affect its compilation regardless,
making it pointless to include in the mega melt.