Update guava: 21.0 -> 27.1-jre
The google-cloud-core component at the managed version requires a
version of Guava newer than 21.0, due to the Maven build mechanism
trying to pull in the Guava BOM at version ${guava.version}.
But version 21.0 of guava does not have guava-bom yet.
So an error like the following occurs:

  [ERROR] Failed to execute goal on project n5-google-cloud: Could not resolve dependencies for project org.janelia.saalfeldlab:n5-google-cloud:jar:3.3.2: Failed to collect dependencies at com.google.cloud:google-cloud-storage:jar:1.108.0 -> com.google.cloud:google-cloud-core:jar:1.93.4: Failed to read artifact descriptor for com.google.cloud:google-cloud-core:jar:1.93.4: Could not find artifact com.google.guava:guava-bom:pom:21.0 in scijava.public (https://maven.scijava.org/content/groups/public) -> [Help 1]

So we are forced to upgrade Guava. I hope the fallout is minimal.