Stop excluding xml-apis
This is a tricky situation. The xml-apis component contains classes that
also ship with the core JRE. However, there can be cases where something
wants to build against a newer version of the XML APIs than is provided
by the version of Java being used to build. For example, when testing
the org.janelia.saalfeldlab:n5-viewer_fiji component with the latest
BOM as of this writing, I encountered the following error:

    [ERROR] sixDimTests(org.janelia.saalfeldlab.n5.bdv.BdvAxisMetadataTest)  Time elapsed: 0.653 s  <<< ERROR!
    java.lang.NoClassDefFoundError: org/w3c/dom/ElementTraversal
	    at org.janelia.saalfeldlab.n5.bdv.BdvAxisMetadataTest.setUp(BdvAxisMetadataTest.java:27)
    Caused by: java.lang.ClassNotFoundException: org.w3c.dom.ElementTraversal
	    at org.janelia.saalfeldlab.n5.bdv.BdvAxisMetadataTest.setUp(BdvAxisMetadataTest.java:27)

And BdvAxisMetadataTest:27 is a call to:

    new org.janelia.saalfeldlab.n5.N5FSReader(...)

Which is part of org.janelia.saalfeldlab:n5

The org.w3c.dom.ElementTraversal class is part of xml-apis v1.4.01,
but not part of the Java 8 JRE. It *is* part of the Java 11 JRE, though.
This leads me to conclude that somehow, n5 or one of its dependencies
was built against a version of xml-apis newer than what Java 8 provides,
and the reference to the ElementTraversal class made it into the
bytecode, and then when classes are being loaded, we see this crash.

Anyway, the problem goes away if we stop excluding xml-apis, because
then everything that wants to build against xml-apis at compile time,
and depend on it at runtime, can do so, rather than relying on the core
JRE XML classes. Although... I believe the core JRE XML classes still
take precedence over the user classpath, so really, it's then relying
on a *mixture* of the core JRE XML classes plus xml-apis, and that
seems like a bad idea to me, but what I do I know.