mega-melt: deal with digit-leading version props
There are a couple of artifacts in pom-scijava that begin with a digit:

* sc.fiji:3D_Blob_Segmentation
* sc.fiji:3D_Objects_Counter
* sc.fiji:3D_Viewer

Prior to d0bf752070d96a2613c42e4e1ab86ebdd07c29ee, these used a
"Fiji_" or "ImageJ_" prefix for their version properties. But this
was standardized to a leading underscore instead, for consistency.

So, for backwards compatibility with older pom-scijava parents, we need
to set: A) the version props with Fiji/ImageJ prefix; and B) the version
props with underscore prefix. Otherwise, these components might not get
their versions pinned properly for builds of older components.