Update components for new major ImgLib2 versions
In testing this upgrade, I needed to write a custom script facilitating
execution of the mega-melt test with some SNAPSHOT components built from
branches. The script I used is preserved in SNAPSHOTS.md here, so that
we can more easily reuse it in the future. This commit also makes a
minor adjustment to the tests/run.sh script allowing the mega-melt to
continue even if the version to be cloned is a SNAPSHOT.

Base components with breaking changes:

* imglib2: 5.12.0 -> 6.1.0
* imglib2-realtransform: 3.1.2 -> 4.0.1

Downstream affected components:

* imagej-deprecated: 0.1.5 -> 0.2.0
* imagej-ops: 1.0.0 -> 2.0.0
* imglib2-algorithm: 0.13.0 -> 0.13.2
* imglib2-algorithm-gpl: 0.3.0 -> 0.3.1
* imglib2-cache: 1.0.0-beta-16 -> 1.0.0-beta-17
* imglib2-imglyb: 1.1.0 -> 2.0.0
* imglib2-roi: 0.13.0 -> 0.14.1
* scifio: 0.44.1 -> 0.45.0
* Colocalisation_Analysis: 3.0.5 -> 3.0.6
* SPIM_Registration: 5.0.24 -> 5.0.25
* z_spacing: 1.1.1 -> 1.1.2
* bigdataviewer-core: 10.4.5 -> 10.4.6
* bigdataviewer-server: 3.1.1 -> 3.1.2
* bigdataviewer-vistools: 1.0.0-beta-31 -> 1.0.0-beta-32
* bigdataviewer_fiji: 6.2.3 -> 6.2.4
* TrackMate: 7.11.0 -> 7.11.1
* n5-imglib2: 4.4.0 -> 5.0.0
* BigStitcher: 1.1.5 -> 1.1.6
* multiview-reconstruction: 1.2.8 -> 2.0.0
* labkit-pixel-classification: 0.1.16 -> 0.1.17
* labkit-ui: 0.3.10 -> 0.3.11

Co-authored-by: Mark Hiner <hiner@wisc.edu>