Update scripting-java: 0.4.1 -> 1.0.0
Yes, it's technically a major version bump, but I happen to know that
this update is only a bug-fix update for newer versions of Java.
But I promoted it to 1.0.0 because it's been stable for a long time,
and I finally released it to Maven Central instead of SciJava Maven.